refactor: API 엔드포인트 복수형 변경
