+++ src/api/groups/ranking.ts
@@ -1,5 +1,5 @@
 import { kyInstance } from "@/api";
-import type { RankingResponse } from "@/api/group/type";
+import type { RankingResponse } from "@/api/groups/type";
 
 export const getAllRanking = async (groupId: number) => {
   const response = await kyInstance
