+++ src/api/solutions/type.ts
