+++ src/view/group/dashboard/NoticeModal/NoticeDetail/index.tsx
@@ -1,5 +1,5 @@
-import type { CommentContent } from "@/api/comment/type";
-import type { NoticeResponse } from "@/api/notice/type";
+import type { CommentContent } from "@/api/comments/type";
+import type { NoticeResponse } from "@/api/notices/type";
 import { IcnClose, IcnEdit, IcnNew } from "@/asset/svg";
 import Avatar from "@/common/component/Avatar";
 import Textarea from "@/common/component/Textarea";
