+++ src/view/group/setting/MemberList/constant.tsx
@@ -1,4 +1,4 @@
-import type { MemberResponse } from "@/api/group/type";
+import type { MemberResponse } from "@/api/groups/type";
 import { IcnCalendarTable } from "@/asset/svg";
 import type { TableDataType } from "@/shared/type/table";
 import {
