+++ src/view/group/solved-detail/CodeSection/CodeHighlighter/rule.ts
@@ -1,4 +1,4 @@
-import type { SolutionLanguage } from "@/api/solution/type";
+import type { SolutionLanguage } from "@/api/solutions/type";
 import Prism from "prismjs";
 
 export const languageMapper: {
