+++ src/view/group/solved-detail/CodeSection/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import type { SolutionLanguage } from "@/api/solution/type";
+import type { SolutionLanguage } from "@/api/solutions/type";
 import CodeHighlighter from "./CodeHighlighter/CodeHighlighter";
 import { sectionWrapper } from "./index.css";
