+++ src/view/user/index/GroupCard/StatusIcon.tsx
@@ -1,4 +1,4 @@
-import type { GroupStatus } from "@/api/group/type";
+import type { GroupStatus } from "@/api/groups/type";
 import { statusIconStyle } from "@/view/user/index/GroupCard/index.css";
 
 const StatusIcon = ({ status }: { status: GroupStatus }) => {
